finalize third segment of home page
